rename dev lobby router
