+++ backend/README
@@ -19,7 +19,7 @@ sudo docker-compose up -d --build
 2. в файлах alembic_/env.py и app/dbManager/dbManager.py в переменных окружения
 задать настройки для подключения к своей базе данных
 4. python migration.py
-5. python app.main.py
+5. uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
